requirements.txt
requests>=2.28.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
google-genai>=0.1.0
psutil>=5.9.0
pyppeteer>=1.0.2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
cryptography>=3.4.0
stripe>=7.0.0
pyautogui>=0.9.54
colorama>=0.4.6

# AI Functionality Dependencies (Optional)
opencv-python>=4.8.0
transformers>=4.30.0
torch>=2.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
numpy>=1.24.0
tensorflow>=2.13.0

# StableMemory Dependencies
sentence-transformers>=2.2.0
aiohttp>=3.8.0
asyncpg>=0.28.0

# Desktop App Dependencies
customtkinter>=5.2.0
tkinter-tooltip>=2.0.0 